Added a delta time cap to avoid missing important physics events
Former-commit-id: e9f14ddda0963537ae6593c2dacc6a877f88c824

+++ common/src/sys/phys.rs
@@ -23,10 +23,10 @@ impl<'a> System<'a> for Sys {
     fn run(&mut self, (terrain, dt, mut positions, mut velocities): Self::SystemData) {
         for (pos, vel) in (&mut positions, &mut velocities).join() {
             // Gravity
-            vel.0.z = (vel.0.z - GRAVITY * dt.0 as f32).max(-50.0);
+            vel.0.z = (vel.0.z - GRAVITY * dt.0).max(-50.0);
 
             // Movement
-            pos.0 += vel.0 * dt.0 as f32;
+            pos.0 += vel.0 * dt.0;
 
             // Basic collision with terrain
             let mut i = 0;
